src/backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
websockets>=12.0

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.13.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.5.2
pydantic-settings>=2.1.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
aiohttp>=3.9.1

# Testing
pytest>=7.4.3
pytest-asyncio>=0.23.2
httpx>=0.25.2

# Development tools
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1
isort>=5.12.0
